modularize team_member
